Adding admin page functionalities and fixing bugs

+++ public/assignment/server/services/user.service.server.js
@@ -9,25 +9,30 @@ module.exports = function(app, userModel) {
 
     // Checks whether the session is authenticated
     var auth = function (req, res, next) {
+        console.log("user.service.server - checking if authorized");
         if (!req.isAuthenticated()) {
+            console.log("user.service.server - Unauthorized");
             res.send(401);
         } else {
+            console.log("user.service.server - Authorize. Passing down the chain.");
             next(); // Callback function given by express to continue in the chain
         }
     };
 
     // Available end points
-    //app.post("/api/assignment/user", createUser);
     app.post("/api/assignment/login", passport.authenticate('local'), login);
     app.post("/api/assignment/register", createUser);
     app.get("/api/assignment/user", getUserByCredentials);
-    app.get("/api/assignment/user", auth, getAllUsers);
     app.get("/api/assignment/user/:id", getUserById);
     app.get("/api/assignment/user", getUserByUsername);
     app.put("/api/assignment/user/:id", auth, updateUserById);
-    app.delete("/api/assignment/user/:id", auth, deleteUserById);
     app.get("/api/assignment/loggedIn", loggedIn);
     app.post("/api/assignment/logout", logout);
+    app.post("/api/assignment/admin/user", auth, createNewUser);
+    app.get("/api/assignment/admin/user", auth, getAllUsers);
+    app.get("/api/assignment/admin/user/:id", auth, getExistingUserById);
+    app.delete("/api/assignment/admin/user/:id", auth, deleteUserById);
+    app.put("/api/assignment/admin/user/:id", auth, updateExistingUserById);
 
     passport.use(new LocalStrategy(localStrategy));
 
@@ -124,15 +129,23 @@ module.exports = function(app, userModel) {
     }
 
     function getAllUsers(req, res) {
-        var userList = userModel.findAllUsers()
-            .then(
-                function (doc) {
-                    res.json(doc);
-                },
-                function (err) {
-                    res.status(400).send(err);
-                }
-            );
+        if (isAdmin(req.user)) {
+            console.log("user.service.server.getAllUsers - Fetching list of all users");
+            userModel
+                .findAllUsers()
+                .then(
+                    function (doc) {
+                        console.log("user.service.server.getAllUsers - result: " + JSON.stringify(doc));
+                        res.json(doc);
+                    },
+                    function (err) {
+                        console.log("user.service.server.getAllUsers - error: " + err.message);
+                        res.status(400).send(err);
+                    }
+                );
+        } else {
+            res.status(403);
+        }
     }
 
     function getUserById(req, res) {
@@ -197,16 +210,20 @@ module.exports = function(app, userModel) {
     }
 
     function deleteUserById(req, res) {
-        var userId = req.params.id;
-        var result = userModel.deleteUserById(userId)
-            .then(
-                function (doc) {
-                    res.json(doc);
-                },
-                function (err) {
-                    res.status(400).send(err);
-                }
-            );
+        if (isAdmin(req.user)) {
+            var userId = req.params.id;
+            var result = userModel.deleteUserById(userId)
+                .then(
+                    function (doc) {
+                        res.json(doc);
+                    },
+                    function (err) {
+                        res.status(400).send(err);
+                    }
+                );
+        } else {
+            res.status(403);
+        }
     }
 
     function loggedIn(req, res) {
@@ -218,4 +235,82 @@ module.exports = function(app, userModel) {
         res.send(200);
     }
 
+    function isAdmin(user) {
+        console.log("user.model - isAdmin - argument passed: " + JSON.stringify(user));
+        return (user.roles.indexOf("admin") != -1);
+    }
+
+    function createNewUser(req, res) {
+        if (isAdmin(req.user)) {
+            var user = req.body;
+
+            userModel
+                .findUserByUsername(user.username) // To check if a user by the same username exists
+                .then(
+                    function (doc) {
+                        if (doc) {
+                            res.json(null);
+                        } else {
+                            return userModel.createUser(user); // Since the user doesn't exist, create one
+                        }
+                    },
+                    function (err) {
+                        res.status(400).send(err);
+                    }
+                )
+                .then(
+                    function (doc) {
+                        return userModel.findAllUsers();
+                    },
+                    function (err) {
+                        res.status(400).send(err);
+                    }
+                )
+                .then(
+                    function (doc) {
+                        res.json(doc);
+                    },
+                    function (err) {
+                        res.status(400).send(err);
+                    }
+                );
+        } else {
+            res.status(403);
+        }
+    }
+
+    function getExistingUserById(req, res) {
+        if (isAdmin(req.user)) {
+            var userId = req.params.id;
+
+            userModel.findUserById(userId).then(
+                function (doc) {
+                    res.json(doc);
+                },
+                function (err) {
+                    res.status(400).send(err);
+                }
+            );
+        } else {
+            res.status(403);
+        }
+    }
+
+    function updateExistingUserById(req, res) {
+        if (isAdmin(req.user)) {
+            var userId = req.params.id;
+            var user = req.body;
+            userModel.updateUser(userId, user)
+                .then(
+                    function (doc) {
+                        res.json(doc);
+                    },
+                    function (err) {
+                        res.status(400).send(err);
+                    }
+                );
+        } else {
+            res.status(403);
+        }
+    }
 };
